src/test/java/com/iyzico/todo/controller/UserControllerTest.java
package com.iyzico.todo.controller;

import com.iyzico.todo.service.UserService;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;
import org.springframework.web.context.WebApplicationContext;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;


/**
 * Created by heval on 23.09.2016.
 */
@RunWith(SpringRunner.class)
@SpringBootTest
public class UserControllerTest {

    private MockMvc mockMvc;

    @MockBean
    private UserService userService;

    @Autowired
    private WebApplicationContext wac;

    @Before
    public void setUp() throws Exception {
        this.mockMvc = MockMvcBuilders.webAppContextSetup(this.wac).build();
    }

    @Test
    public void testAddUser() throws Exception {
        mockMvc.perform(post("/addUser"))
                .andExpect(status().isOk())
                .andExpect(view().name("index"))
                .andExpect(handler().handlerType(UserController.class))
                .andExpect(handler().methodName("addUser"))
                .andExpect(model().attribute("success", true))
                .andReturn();
    }
}
